使用mybatis generator重构

+++ src/main/java/cn/fengylb/mycommunity/mycommunity/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Long id,
+            @RequestParam(value = "id", required = false) Integer id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
@@ -62,7 +62,7 @@ public class PublishController {
         question.setTitle(title);
         question.setDescription(description);
         question.setTag(tag);
-        question.setCreator(Long.valueOf(user.getId()));
+        question.setCreator(user.getId());
         question.setAccountId(user.getAccountId());
         questionService.createOrUpdate(question);
         return "redirect:/";
